% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.plotLift.R
\name{h2o.plotLift}
\alias{h2o.plotLift}
\title{Plot the Lift Curve of one or more model}
\usage{
h2o.plotLift(models, newdata = NULL, xval = FALSE, groups = 100)
}
\arguments{
\item{models}{list of h2o models class H2OBinomialModel}

\item{newdata}{dataframe class H2OFrame}

\item{xval}{if TRUE plot the Lift Curves on cross validation}
}
\value{
ggplot graph
}
\description{
Plot the Lift Curve of one or more model
}
\examples{

}
\seealso{
h2o.plotROC(), h2o.plotVarImp()
}
\author{
Elías Alegría <elias.alegria@ug.uchile.cl>
}
