% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.plotROC.R
\name{h2o.plotROC}
\alias{h2o.plotROC}
\title{Plot the ROC Curves}
\usage{
h2o.plotROC(models, newdata = NULL, xval = FALSE)
}
\arguments{
\item{models}{list of h2o models class H2OBinomialModel}

\item{newdata}{dataframe class H2OFrame}

\item{xval}{if TRUE plot the ROC Curves on cross validation}
}
\value{
ggplot graph
}
\description{
Plot the ROC Curves
}
\examples{

}
\seealso{
h2o.plotLift(), h2o.plotVarImp()
}
\author{
Elías Alegría <elias.alegria@ug.uchile.cl>
}
